Eliminada carpeta Brazo adentro de Brazo. Otro proyecto entero, ni idea de como llegó ahí.
Actualizado valor de constante ALPHA en el filtro de movimiento a 0.7
Actualizadas impresiones en pantalla oled.
Agregadas impresiones por UART para posición del brazo utilizando mismos sprintf().
Corregida función decreaseAngle() para que retorne sin modificaciones para valores de ángulo iguales o menores a 0 (antes >= 180, error de copy paste).
Cambiados vectores de acelerómetros y giroscopios para ser SIGNED INTS (antes unsigned).
Cambiado timestamp: uint32_t -> uint16_t.

+++ Brazo/Core/Inc/BrazoFSM1.h
@@ -19,10 +19,10 @@
 #include "ssd1306_fonts.h"
 
 typedef struct {
-  uint16_t acelerometros[2][3];
-  uint16_t giroscopios[2][3];
+  int16_t acelerometros[2][3];
+  int16_t giroscopios[2][3];
   uint8_t pote;
-  uint32_t timeStamp;
+  uint16_t timeStamp;
   uint8_t angle;
   uint8_t channel;
   uint8_t flag;
